pyproject.toml
[tool.poetry]
name = "navanchauhan.github.io"
version = "0.1.0"
description = ""
authors = ["navanchauhan <navanchauhan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Jinja2 = "^3.1.2"
markdown2 = "^2.4.10"
Pygments = "^2.16.1"
"Mastodon.py" = "^1.8.0"
Pillow = "10.0.1"
requests = "^2.31.0"
certifi = "^2023.7.22"
urllib3 = "^2.0.7"
latex2mathml = "^3.77.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
